Fix issues with incorrect date display in reports.
The date filter in the report client incorrectly calls startOf / endOf
multiple times when building the query parameters from a date selected
in the UI. This can lead to the filter becoming incorrect when repeated
applications of the offset  cause the date to either move forward or
backward, across a date boundary.

For example, if the browser timezone is UTC, and the profile timezone is
+6:

- User selects 2015-09-07 as the end date
- First endOf call changes this to 2015-09-07T24:00:00 (technically naive, but
  treated by Moment as a datetime in the browser timezone)
- Moment tz call applies the +6 making it 2015-09-08T06:00:00
- endOf applied again, end date is now 2015-09-08T24:00:00

The inverse can happen with startOf and a negative offset timezone as well.

Another issue is that the keys in the daystoshow hashmap had the profile
timezone applied. This causes issues with functions inside the loadData
block that build their own API calls. Profile timezone application could
happen multiple times resulting in incorrect calls.

Additionally, the keys in the daystoshow hashmap are used by the report
plugins themselves. These plugins use the localeDate / localeDateTime
util functions to display the date. These functions would again attempt
to apply to profile timezone, causing issues with rollover due to being
applied multiple times. This was particularly problematic with callers of
localeDate as the day of the week was passed to translation before final
timezone application. Additionally, by using toLocaleDateString /
toLocaleTimeString functions the browser timezone was being applied as
well, introducing another vector for date rollover issues.

This branch changes all display functions to display in the browser
timezone, makes all API requests use the current profile timezone where
appropriate, and ensures timezones offsets are not repeatedly applied.

I believe this fixes issues #8219, #8176, #6853, and #5659.